+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [NEXT_RELEASE]
+### Changed
+- Back to Godep for dependencies management
+
 ## [0.11.0] - 2017-12-11
 ### Added
 - [helm] support rbac