+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Back to Godep for dependencies management
 - Update `client-go` lib to version 4.0
+### Added
+- alias `app log` to `app logs`
+- `app logs` now  support shorthand `-f` to `--follow` and `-n` to `--lines`  
 
 ## [0.11.0] - 2017-12-11
 ### Added