+++ CHANGELOG.md
@@ -6,9 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Back to Godep for dependencies management
 - Update `client-go` lib to version 4.0
+
 ### Added
 - alias `app log` to `app logs`
 - `app logs` now  support shorthand `-f` to `--follow` and `-n` to `--lines`  
+
+### Fixed
+- Env vars being set or unset for the app on deploy update errors
 
 ## [0.11.0] - 2017-12-11
 ### Added